+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - Fixed parsing bug when there is no comment in an OpenSSH public key.
 - Made members of BinaryRange private.
 - Added GenericPublicKey for other key-types in the OpenSSH Public Key format.
+- Added coverage and some more unit tests.
 
 ## 0.7.1